# Speed up some `is_ascii*` methods on `char` and `u8`; add exhaustive tests for all `is_ascii*` methods