--- a/library/core/src/num/mod.rs
+++ b/library/core/src/num/mod.rs
@@ -258,6 +258,41 @@ impl isize {
 /// If 6th bit set ascii is upper case.
 const ASCII_CASE_MASK: u8 = 0b0010_0000;
 
+// This macro creates part of a function that handles up to eight strips of consecutive
+// matching codepoints. The strips must all be in separate 32-codepoint chunks
+// (codepoints 0 to 31, 32 to 63, 64 to 95, 96 to 127, 128 to 159, 160 to 191,
+// 192 to 223, or 224 to 255).
+#[allow_internal_unstable(const_slice_index)]
+macro_rules! handle_strip_of_each_chunk {
+    ($x: ident, $starting_codepoints: expr, $strip_lengths: expr) => {{
+        // 32-codepoint chunk number.
+        //
+        // SAFETY: Guaranteed by bit operations to be in `0..8`.
+        let chunk_number = ($x as u8 & 0b1110_0000).wrapping_shr(5) as usize;
+
+        // `const` to type check and to ensure all element evaluations are done at
+        // compile time
+        const STARTING_CODEPOINTS: [u8; 8] = $starting_codepoints;
+        // Subtract the starting codepoint of this chunk from the input codepoint. This
+        // will make sure that the matching codepoints in this strip are in
+        // `0..length_of_strip`.
+        //
+        // SAFETY: The array length is 8 and `chunk_number` is guaranteed to be in
+        // `0..8`.
+        let x = $x.wrapping_sub(*unsafe { STARTING_CODEPOINTS.get_unchecked(chunk_number) });
+
+        // `const` to type check and to ensure all element evaluations are done at
+        // compile time
+        const STRIP_LENGTHS: [u8; 8] = $strip_lengths;
+        // Check whether the adjusted value of the input codepoint is in
+        // `0..length_of_strip`.
+        //
+        // SAFETY: The array length is 8 and `chunk_number` is guaranteed to be in
+        // `0..8`.
+        x < *unsafe { STRIP_LENGTHS.get_unchecked(chunk_number) }
+    }};
+}
+
 impl u8 {
     uint_impl! { u8, u8, i8, NonZeroU8, 8, 255, 2, "0x82", "0xa", "0x12", "0x12", "0x48", "[0x12]",
     "[0x12]", "", "", "" }
@@ -545,7 +580,12 @@ impl u8 {
     #[rustc_const_stable(feature = "const_ascii_ctype_on_intrinsics", since = "1.47.0")]
     #[inline]
     pub const fn is_ascii_alphanumeric(&self) -> bool {
-        matches!(*self, b'0'..=b'9' | b'A'..=b'Z' | b'a'..=b'z')
+        let x = *self;
+        handle_strip_of_each_chunk!(
+            x,
+            [0, b'0', b'A', b'a', 0, 0, 0, 0],
+            [0, 10, 26, 26, 0, 0, 0, 0]
+        )
     }
 
     /// Checks if the value is an ASCII decimal digit:
@@ -616,7 +656,8 @@ impl u8 {
     #[rustc_const_stable(feature = "const_ascii_ctype_on_intrinsics", since = "1.47.0")]
     #[inline]
     pub const fn is_ascii_hexdigit(&self) -> bool {
-        matches!(*self, b'0'..=b'9' | b'A'..=b'F' | b'a'..=b'f')
+        let x = *self;
+        handle_strip_of_each_chunk!(x, [0, b'0', b'A', b'a', 0, 0, 0, 0], [0, 10, 6, 6, 0, 0, 0, 0])
     }
 
     /// Checks if the value is an ASCII punctuation character:
@@ -654,7 +695,15 @@ impl u8 {
     #[rustc_const_stable(feature = "const_ascii_ctype_on_intrinsics", since = "1.47.0")]
     #[inline]
     pub const fn is_ascii_punctuation(&self) -> bool {
-        matches!(*self, b'!'..=b'/' | b':'..=b'@' | b'['..=b'`' | b'{'..=b'~')
+        // Add 6 to the codepoint so that each strip of consecutive matching codepoints
+        // is in a separate 32-codepoint chunk. For more details, see the comment above
+        // the `handle_strip_of_each_chunk` macro definition.
+        let x = (*self).wrapping_add(6);
+        handle_strip_of_each_chunk!(
+            x,
+            [0, b'!' + 6, b':' + 6, b'[' + 6, b'{' + 6, 0, 0, 0],
+            [0, 15, 7, 6, 4, 0, 0, 0]
+        )
     }
 
     /// Checks if the value is an ASCII graphic character:
@@ -739,7 +788,13 @@ impl u8 {
     #[rustc_const_stable(feature = "const_ascii_ctype_on_intrinsics", since = "1.47.0")]
     #[inline]
     pub const fn is_ascii_whitespace(&self) -> bool {
-        matches!(*self, b'\t' | b'\n' | b'\x0C' | b'\r' | b' ')
+        // The long binary number has bit indexes starting at 0 on the right and going
+        // leftward until it ends at bit index 32. The bit index corresponds to the
+        // codepoint of the input `u8`. The value of the bit there is 1 iff the `u8` is
+        // an ASCII whitespace codepoint.
+        let x = *self;
+        x <= b' '
+            && (0b1_0000_0000_0000_0000_0011_0110_0000_0000_u64.wrapping_shr(x as u32) & 1) != 0
     }
 
     /// Checks if the value is an ASCII control character: