--- a/library/core/src/num/mod.rs
+++ b/library/core/src/num/mod.rs
@@ -689,7 +689,7 @@ impl u8 {
         // Add 6 to the codepoint so that each strip of consecutive matching codepoints
         // is in a separate 32-codepoint chunk. For more details, see the comment above
         // the `handle_strip_of_each_chunk` macro definition.
-        let x = (*self).wrapping_add(6);
+        let x = self.wrapping_add(6);
         handle_strip_of_each_chunk!(
             x,
             [0, b'!' + 6, b':' + 6, b'[' + 6, b'{' + 6, 0, 0, 0],
@@ -784,8 +784,7 @@ impl u8 {
         // codepoint of the input `u8`. The value of the bit there is 1 iff the `u8` is
         // an ASCII whitespace codepoint.
         let x = *self;
-        x <= b' '
-            && (0b1_0000_0000_0000_0000_0011_0110_0000_0000_u64.wrapping_shr(x as u32) & 1) != 0
+        x <= b' ' && ((0b1_0000_0000_0000_0000_0011_0110_0000_0000_u64 >> x) & 1) != 0
     }
 
     /// Checks if the value is an ASCII control character: