@ChaiTRex 2022-08-15
> > I've also exhaustively tested that all valid char and u8 inputs provide the exact same outputs as the current versions, so, except for invalid char values, if the current version works, so does the new version.
> 
> We should maybe add tests like that? Depends how slow it is. But the cost of running 2 functions on every u8 is nothing, and for every char it's what, only 1 million? That's basically nothing, so I don't expect it to cause perf issues in running tests (except under miri, you may need to run a subset of the tests under miri). (invalid chars are UB anyways, you don't need to consider them at all).
> 
> Like, have a model "obviously correct" function for all `is_ascii_*` functions (Probably just a range check/range checks), then loop through Every `u8` and Every `char` and assert that they give the same answer. As a sketch, maybe something like:
> 
> ```rust
> for value in 0..=u8::MAX {
>     let uppercase = (b'A'..=b'Z').contains(&value);
>     let lowercase = (b'a'..=b'z').contains(&value);
>     let numeric = (b'0'..=b'9').contains(&value);
>     let control = (0..=0x1f).contains(&value) || value == 0x7f;
>     let graphic = (b'!'..=b'~').contains(&value);
>     let hexdigit = numeric | (b'A'..=b'F').contains(&value) | (b'a'..=b'f').contains(&value);
>                                                                                               
>     assert_eq!(value.is_ascii(), (0..128).contains(&value));
>     assert_eq!(value.is_ascii_alphabetic(), uppercase | lowercase);
>     assert_eq!(value.is_ascii_alphanumeric(), uppercase | lowercase | numeric);
>     assert_eq!(value.is_ascii_control(), control);
>     assert_eq!(value.is_ascii_digit(), numeric);
>     assert_eq!(value.is_ascii_graphic(), graphic);
>     assert_eq!(value.is_ascii_hexdigit(), hexdigit);
>     assert_eq!(value.is_ascii_lowercase(), lowercase);
>     assert_eq!(value.is_ascii_uppercase(), uppercase);
> }
> ```
> 
> The thing I'm worried about is breaking some edge case and, because we don't have full test coverage, nothing actually notices.
> 
> For example, try adding a `| 0xfd` to the current `is_ascii_whitespace`, and then running the full set of `core` tests! Nothing notices that. Maybe some test somewhere would notice, but we're definitely not _intentionally_ testing for it.
> 
> It's not too difficult to have full exhaustive tests here, so we should, especially if we're making the implementation more complex.

I've now added tests for the `is_ascii*` methods for `char` and `u8` based on the current 1.63.0 definitions.