@ChaiTRex 2022-08-16
* Requested changes by @erikdesjardins worked nicely, so they're put in commit 3.
* Made `is_ascii` method tests more understandable by switching to `matches!`.