@ChaiTRex 2022-08-16
* Finished simplification of macros in commits 5 and 7.
* Reverted `char::is_ascii_alphabetic` to original version because LLVM 15 now optimizes it to be just as fast as my changed version, as @erikdesjardins pointed out, in commit 6.